refactor(testing): hoist install_fake_plugin fixture to tests/conftest.py
Move ``fake_vst3_plugin`` and ``install_fake_plugin`` from
``tests/data/vst/conftest.py`` to ``tests/conftest.py`` so they are
discoverable from sibling test trees (``tests/tools/``, ``tests/evaluation/``,
``tests/integration/``). ``tests/data/vst/_fake_plugin.py`` stays put — only
the fixtures' discovery scope widens. The old VST-scoped conftest had no
other content, so the file is removed.

Wave 1 of the rollout in #1258. Unblocks the cross-tree fake-plugin twins
and the consolidation of ad-hoc plugin doubles tracked under that issue.

Refs #1258
